server compiles with shared
