+++ shared/index.ts
@@ -1,2 +1,3 @@
 // Main entry point for shared CRDT library
 export * from "./crdt/database/CRDTDatabase";
+export * from "./crdt/PixelDataCRDT";
